pipeline/requirements.txt
pymysql
python-dotenv
openai
selenium
